Boston Review: Fix broken translator; remove FW code
The translator no longer worked with the current site. The fix is
implemented using the EM translator, and the old FW code is completely
removed.

Test cases are updated and new tests are added for multiple authors and
special projects.

Note that the ISSN field is no longer added to all translated items
because in general there's no way to determine if an online article
belongs to a particular print issue under that ISSN. Further advices is
appreciated for this particular issue.

See also #3105 for the FW problem.