Fix wikidata translator bug
Fix wikidata translator author bug,
which failed when it encountered a "null"
creator.

This occurs when the wikidata item
it a redirect.

If no value is found, continue to the next
iteration of the loop. This will skip
adding the item instead of throwing an error.

TODO: Fix duplicate authors, and
get labels for redirected items instead of
skipping them.