DOI web translator: Offer the current page as a choice among the multiple
If the DOI translator turns out to be the one that matches the current
page, and if the DOI (this) translator detects the item type as
"multiple", offer the current page as one of the choices (indeed, the
first choice).

In the item selection dialog, this special choice will show up as
"Current Webpage", followed by the page title in parenthesis if the
title is present.

The specific situation that triggers this behavior is the following

1. None other than the DOI web translator (lowest priority) matches the
   current page
2. The current page itself is not the main resource identified by DOI
   present on the current page (e.g. a page with its own DOI in URL)

Resolves #3126.