DOI.js: Don't offer current page if it may be a duplicate among the multiple
Use a dissimilarity metric based on item URL and title to de-duplicate
the special "current web page" item. A test case is added for this
scenario.

Also, add a visual cue (text "possibly current web page" in parentheses
after the title) in the item selection dialog for the choice that may be
referring to the current web page.