DOI.js: Revise handling of current-page in multiple for Connector update
In anticipation of updates to the Connector in the item-selector, where
a special key -- the translation-initiating page's URL -- will be used
to label it as the current page with proper localization, changes
are made here to take advantage of that upcoming feature, and display
non-localized title label as fallback before the Connector is updated.

(Cf. https://github.com/zotero/translators/pull/3130#issuecomment-1717803990)

The logic of doing so is a bit non-trivial, and the code is now more
extensively commented.

Some minor simplifications are done in passing.