Frontiers journals: Implement saving multiple from search page
Multiple-saving now works on the search page in addition to the article
listings by journal or topic.

This is done by getting the numerical articleID from the search page,
from which the article URL can be constructed (as /articles/articleID).
This URL will be redirected to the proper, DOI-based URL.