IEEE Xplore: Don't use export format; save full text PDF
The "endpoint" for export formats (BibTeX, RIS, etc.) may be unreliable
unless the referrer is sent, but this is not yet feasible for
Chromium-family browsers in Connector.

(See: https://forums.zotero.org/discussion/108011/ieee-xplore-falied)

This is changed to a scraper that looks at just the page content,
therefore avoiding the endpoint call.

The pages are rendered on the client side, so we read the data source as
JSON directly, without scraping the DOM.

In addition, the following changes are made --

- The full-text PDF file is now saved as an attachment (if the user has
  access to it).
- The translator now also works on IEEE Xplore's own full-text PDF
  viewer web app. This is done via an additional network request to the
  abstract page where the JSON metadata are hosted. A test case for this
  is added.
- Tags (keywords) are deduplicated.
- Only the electronic-media ISBN and ISSN are kept.
- Other minor bugfixes.