NASA ADS (web translator): Update logic for type detection (for UI only)
Based on the bibcode spec, the type returned by detectWeb() is made more
accurate. This results in more intuitive icon display by the Connector.

Note that the translated item's actual type is not affected by this;
it is instead handled by 'ADS Bibcode.js' search translator, which uses
both Bibcode and the API-returned RIS data to determine the type, with
greater accuracy.