VuFind: Don't exhaustively try formats for each item
Determine the available set of export formats (i.e. input formats for
this translator) from the web page initiating the translation, choose
the first supported one, and use it for scraping all of the multiple
items. By doing this we don't have to repeat the trial-and-error for
input formats every single time.