VuFind: New logic for IDing input format; update multiple selector
Not all VuFind sites advertise all of the supported export formats. For
instance, a site may support MARC but stop short of offering a link for
MARC export.

Therefore we try something very similar to the original (#2969) PR.
Namely, we start with an initial format defined by ourselves independent
of what the site says. This initial format can be set manually or
looked up from a small list of known sites.

If the initial one doesn't work, we consult the list of formats offered
by the site, and use those formats as fallback, trying one after
another.

The difference from #2969 is that we don't repeat this trial-and-error
for every single input of the multiple. Instead, if we find a supported
format, we use it for subsequent items.

If the input format is set manually in the `exports`, don't do fallback.

In addition, we add some more selectors for the itemType-identifying
<span> elements for detectWeb()