Add translator for Libris
Libris is a service by The National Library of Sweden ("Kungliga
biblioteket"). It collects information from several Swedish
libraries. For more information see
http://libris.kb.se/help/about_libris_eng.jsp?language=en.

This translator is based on the translator for National Library of
Poland ISBN. Based on the discussion in #3036 it filters only books
published in Sweden. It uses the Xsearch API:
http://libris.kb.se/help/xsearch_eng.jsp?language=en.