Twitter: Quick fix for translation error
We've lost URL unshortening on most versions of the site for now, but at
least it doesn't error. They seem to be A/B testing, or maybe just
randomizing, the way links are displayed.

Sometimes (and *always* with an ad blocker enabled) I'm actually seeing
links displayed in full on the page, without t.co shortening; without an
ad blocker, it varies between that layout and one where the full URLs
are nowhere in the page source, meaning there's likely no way for us to
unshorten them.

https://forums.zotero.org/discussion/114434/twitter-x-translator-extension-resorts-to-generic-embedded-metadata