Improve citation of archived urls from Wayback
People citing Wayback urls presumably want the metadata
from the website or news article itself.

Using the embedded metadata translator improves these results
i.e. including the author, and the date the original article was
published instead of the archive date.

Ideally we would actually want to use the translators specific
to the particular url itself but this is a good first pass
that improves things.

This also sets the access date to the date the page was
archived since this is the date the metadata is being
extracted from.