ePrint: fix for MathJax lazy typesetting affecting tests
Running tests for the PS format addition change revealed that ePrint pages
seem to have switched to using MathJax lazy typesetting. That leads to
differences in expected vs scraped abstract text when run under Scaffold,
but not when running manually from browser. Fix by pulling the abstract
from the OG tag if MathJax elements are not typeset yet.

Kudos to @AbeJellinek who found the root cause and suggested the workaround.

Signed-off-by: Alex T. <frozen.and.blue@gmail.com>
Co-authored-by: Abe Jellinek <jellinek@berkeley.edu>