DOI Content Negotiation: Handle upcoming Crossref outage
Crossref is retrying their maintenance from last month on Sep. 16-17,
which means content negotiation for Crossref DOIs will be down for
another 24-hour period. (Last time they didn't actually realize it was
down -- I told them they were returning 200s with 503 error messages in
the body -- and this time they're doing it intentionally...) So during
that period, make an extra request to check the registration agency of
each DOI, and if it's Crossref, use their REST API, which is remaining
up.