Store reply summary instead
The Bsky API in fact returns all replies, which is a non-starter for popular posts which may have hundreds or thousands – we don't want to store a separate note for each. Instead, we now store a note with the number of direct replies at the time of storage, as an ISO datetime.