Update EUR-Lex.js
This commit improves the import of files as statutes, i.e. EU regulations. The "bill" file type in Zotero already has a field named "Legislative Body". The file type 'statute' does not. When importing a file of this type, the translator writes the file type in the 'Extra' field. You cannot have unofficial variables in the extra field if you want your style to be registered in the official Zotero style repository. The solution is to use the recognised file variables. The correct CSL field mapping for 'Legislative Body' is 'authority', so 'authority' must be imported into the 'Extra' field for statutes.