ACM Queue Translator
Adds a traslator for https://queue.acm.org/.
First, try to resolve a DOI, which works for most
older articled. If it fails, mainly for more recent
articles, fetch metadata to the best of our ability
from the DOM of the article page and the page of the
journal issue.

Fixes #3496